New York City Taxi Trip/requirements.txt
streamlit==0.86.0
seaborn==0.11.1
pandas==1.1.3
matplotlib==3.3.2
Pillow==8.3.1
scikit_learn==0.24.2
